+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Handle selected overdue appointments state in overdue screen
 - Add support for selected overdue appointments download
 - Show selected overdue appointments count in overdue screen
+- Show download and share button in overdue search screen when search results are loaded
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes